Fix cards clipping when played by oppo
Sometimes when the oppo played, the cards were too low and would clip
through the player's played card. We now add a Z offset to the oppo
played cards so that they come from a bit higher.

The reason this happened is because of the play on perspective. The oppo
hand is almost level with the war pile, so it doesn't have much Z offset
when transitioning the card from the hand to the pile. Adding the Z
offset when the card is played is a workaround.

This fixes #8