Fix inconsistent hover glow color
All the glow colors share the same material, meaning that it's important
to set the correct color of the glow when changing cards. Turns out
sometimes the glow color of a card that stops being hovered gets set (at
the same time as when the visibility of the glow effects is turned off)
after the moment a newly visible glow has its color set.

On top of that, the handling of the hovering state of a card dodgy. The
state is now not-duplicated; We also disable hovering other cards
whenever we are dragging one.

This stops the color from being set when the visibility is turned off.

This fixes #7