Fix cards spazzing out when cursor between two
Fixes #6. Though the placement is still wonky, at least the flickering
is gone.

Thanks to rotating the cards to be aligned with the camera (this way,
when changing the height of the card, it doesn't change which card is
in front).

We also add an "underlay" so that it's not possible to hover over a card
that is behind another card in normal position. Otherwise, the hover
would swap card endlessly.