Remove debug scene
Now that for some unknown reason the default scene loads super fast,
there is no reason to keep around the debug scene. On top of that,
having two distinct scenes between debug and release might lead to
desynced values.

For that matter, the orentation of the player pile differed in the debug
and release scene, causing lot of confusion when displaying the player
scores.